changed status to long

+++ core/src/main/scala/com/jabong/dap/common/udf/UdfUtils.scala
@@ -528,7 +528,7 @@ object UdfUtils {
     if (null != s && (s.contains(DataSets.WINDOWS) || s.contains(DataSets.ANDROID) | s.contains(DataSets.IOS))) s else s1
   }
 
-  def successOrder(i: Int): Int = {
+  def successOrder(i: Long): Int = {
     val successCodes = Array(3, 4, 5, 6, 7, 11, 17, 24, 33, 34)
     if (successCodes.contains(i)) {
       return 1
